# Conflicts:
#	src/onegov/org/locale/de_CH/LC_MESSAGES/onegov.org.po
#	src/onegov/org/locale/fr_CH/LC_MESSAGES/onegov.org.po
#	src/onegov/org/locale/it_CH/LC_MESSAGES/onegov.org.po

+++ src/onegov/org/models/ticket.py
@@ -101,6 +101,10 @@ def refresh_submission_invoice_items(
         request.session.add(invoice)
         self.ticket.invoice = invoice
 
+    # update the invoicing party
+    if self.submission is not None:
+        invoice.invoicing_party = self.submission.invoicing_party
+
     old_items = sorted(invoice.items, key=attrgetter('group'))
     new_items: list[InvoiceItem] = []
     unused: set[InvoiceItem] = set(old_items)
@@ -830,7 +834,23 @@ class ReservationHandler(Handler):
         return self.resource.reply_to
 
     def prepare_delete_ticket(self) -> None:
-        for reservation in self.reservations or ():
+        from libres.db.models import ReservedSlot
+
+        reservations = self.reservations
+        if not reservations:
+            return
+
+        # FIXME: replace this with scheduler.remove_reservation
+        # all reservations of a ticket share the token; delete their reserved
+        # slots too, otherwise deleting the reservations directly (instead of
+        # via the scheduler) leaves orphaned slots blocking the calendar
+        token = reservations[0].token
+        self.session.query(ReservedSlot).filter(
+            ReservedSlot.reservation_token == token,
+            ReservedSlot.source_type == 'reservation',
+        ).delete(synchronize_session=False)
+
+        for reservation in reservations:
             self.session.delete(reservation)
 
     @cached_property
